requirements.txt
mysql-connector-python>=8.0.0
tkcalendar>=1.6.1
pillow>=8.0.0
bcrypt>=3.2.0
python-dotenv>=0.19.0
pytest>=6.0.0



TO run the application, use the following command:
```bash
$env:PYTHONPATH = 'd:\HARITH\Parsenal\.NIT_HND\Unit 27 - Advanced Programming\Assignment\Development\System'; & "D:/HARITH/Parsenal/.NIT_HND/Unit 27 - Advanced Programming/Assignment/.venv/Scripts/python.exe" -m hospital_management_system.gui.main_app
